добавлен загрузчик комментариев
