+++ js/picture.js
@@ -1,10 +1,67 @@
-import { initCommentList, renderComments } from './comment.js';
+const COMMENTS_COUNT_SHOWN = 5;
 
 const bigPictureElement = document.querySelector('.big-picture');
+
+const commentsListElement = bigPictureElement.querySelector('.social__comments');
+const commentCountElement = bigPictureElement.querySelector('.social__comment-count');
+const totalCommentCountElement = bigPictureElement.querySelector('.social__comment-total-count');
+const commentsLoaderElement = bigPictureElement.querySelector('.comments-loader');
+
+const commentElement = document
+  .querySelector('#comment')
+  .content.
+  querySelector('.social__comment');
+
+let commentsCountShown = 0;
+let comments = [];
+
+
+const createComment = ({avatar, message, name}) => {
+  const newComment = commentElement.cloneNode(true);
+  newComment.querySelector('.social__picture').src = avatar;
+  newComment.querySelector('.social__picture').alt = name;
+  newComment.querySelector('.social__text').textContent = message;
+
+  return newComment;
+
+};
+
+const renderComments = () => {
+  commentsCountShown += COMMENTS_COUNT_SHOWN;
+
+  if (commentsCountShown >= comments.length) {
+    commentsLoaderElement.classList.add('hidden');
+    commentsCountShown = comments.length;
+  } else {
+    commentsLoaderElement.classList.remove('hidden');
+  }
+
+  const fragment = document.createDocumentFragment();
+  for (let i = 0; i < commentsCountShown; i++) {
+    const comment = createComment(comments[i]);
+    fragment.append(comment);
+  }
+
+  commentsListElement.innerHTML = '';
+  commentsListElement.append(fragment);
+
+  commentCountElement.textContent = commentsCountShown;
+  totalCommentCountElement.textContent = comments.length;
+
+};
+
+// const initCommentList = () => {
+//   commentCountElement.classList.add('hidden');
+//   commentsLoaderElement.classList.add('hidden');
+// };
+
+const onCommentsLoaderClick = () => renderComments();
+
 const bodyElement = document.querySelector('body');
 const closePictureButtonElement = bigPictureElement.querySelector('.big-picture__cancel');
 
 const hidePicture = () => {
+  commentsCountShown = 0;
   bigPictureElement.classList.add('hidden');
   bodyElement.classList.remove('modal-open');
   document.addEventListener('keydown', onDocumentKeydown);
@@ -35,13 +92,19 @@ const showPicture = (pictureData) => {
   bodyElement.classList.add('modal-open');
   document.addEventListener('keydown', onDocumentKeydown);
 
-  renderComments(pictureData.comments);
-  initCommentList();
+  comments = pictureData.comments;
+  if (comments.length > 0) {
+    renderComments();
+  }
+
+
+  // initCommentList();
 
   renderPicture(pictureData);
 
 };
 
 closePictureButtonElement.addEventListener('click', onClosePictureButtonClick);
+commentsLoaderElement.addEventListener('click', onCommentsLoaderClick);
 
 export {showPicture};
